remove excess flex props, rename to Flex and FlexItem, Box component will be separate
