+++ packages/gamut-labs/src/index.ts
@@ -1,2 +1,2 @@
-export * from './Box';
+export * from './Flex';
 export * from './brand';
