--- packages/styleguide/stories/Core/Layouts/Box.stories.mdx
@@ -1,26 +1,0 @@
-import { Meta, Props, Canvas, Story } from '@storybook/addon-docs/dist/blocks';
-import { StoryStatus } from '~styleguide/blocks';
-import { Box } from '@codecademy/gamut-labs/src';
-import { Box as BoxElement, defaultGridProps } from './Elements';
-import { select, boolean, number } from '@storybook/addon-knobs';
-
-<Meta title="Layouts/Box (Labs)" component={Box} />
-
-# Container
-
-<StoryStatus status="stable" />
-
-A flex box primitive that allows you create markup configurable flex containers
-
-<Canvas>
-  <Story name="Box" args={{}}>
-    {(args) => (
-      <Box {...args}>
-        <BoxElement>Element 1</BoxElement>
-        <BoxElement>Element 2</BoxElement>
-      </Box>
-    )}
-  </Story>
-</Canvas>
-
-<Props story="Box" />
