+++ packages/styleguide/stories/Core/Layouts/Flex.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Props, Canvas, Story } from '@storybook/addon-docs/dist/blocks';
 import { StoryStatus } from '~styleguide/blocks';
-import { Flex } from '@codecademy/gamut-labs/src';
+import { Flex, FlexItem } from '@codecademy/gamut-labs/src';
 import { Box as BoxElement, defaultGridProps } from './Elements';
 import { select, boolean, number } from '@storybook/addon-knobs';
